landing-page/pyproject.toml
[tool.poetry]
name = "landing-page"
version = "0.1.0"
description = ""
authors = ["en-GB <67548055+en-GB@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
markdown-it-py = {extras = ["plugins"], version = "^2.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
